requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
Appium-Python-Client==2.6.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==37.0.3
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
Faker==8.10.3
fett==0.3.2
h11==0.13.0
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
networkx==2.5.1
outcome==1.2.0
packaging==20.9
phpserialize==1.3
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.21
Pygments==2.9.0
pymongo==3.11.4
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.2
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rstcheck==3.3.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
snooty-lextudio==1.11.6
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
text-unidecode==1.3
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.5
watchdog==1.0.2
wsproto==1.1.0
yarl==1.6.3
